Timmy: Added in what I believe to be the most sensical implementation of TCA to all the necessary features to have our SecurityApp fully working.
There are revision notes within The MapFeature, and various comments that I believe we need to discuss before commencing forward with building this out. It's fairly complex, so as to avoid further refactoring, we need to decide on how to move forward within integrating the mapFeature as well as the ServiceAccept feature.

This will greatly improve how we implement that various mapFeatures in the Athena User app, once we figure how best to implement this in the Security App.